+++ src/app/interfaces/personal-project.interface.ts
@@ -1,0 +1,9 @@
+export interface PersonalProjects {
+  personalProjects: PersonalProject[];
+}
+
+export interface PersonalProject {
+  name: string;
+  code: string;
+  app?: string;
+}