+++ src/app/shared/main/main.component.ts
@@ -6,6 +6,7 @@ import { TrainingComponent } from './components/training/training.component';
 import { LanguagesComponent } from './components/languages/languages.component';
 import { WorkExperiencesComponent } from './components/work-experience/work-experiences.component';
 import { TagsComponent } from './components/tags/tags.component';
+import { PortfolioProjectsComponent } from './components/portfolio-projects/portfolio-projects.component';
 
 @Component({
   selector: 'app-main',
@@ -16,6 +17,7 @@ import { TagsComponent } from './components/tags/tags.component';
     WorkExperiencesComponent,
     LanguagesComponent,
     TagsComponent,
+    PortfolioProjectsComponent,
   ],
   templateUrl: './main.component.html',
   styleUrl: './main.component.scss',